% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat.R
\name{cheat}
\alias{cheat}
\title{Prints out the code that is the solution for exercises in Q3.1}
\usage{
cheat(exercise_n)
}
\arguments{
\item{exercise_n}{numeric variable}
}
\description{
Prints out the code that is the solution for exercises in Q3.1
}
